(#3369) - disable blobSupport in Chrome, fix Travis
This essentially reverts fce1c97, meaning we're back to
using the 404 test to disable blob support in Chrome/Android/Opera.

This is somewhat flakey because we're using one bug to detect
another, but I'm really hoping to avoid UA sniffing here.

I also moved `idb-blob-support.js` to its own file and changed
the Travis tests so that we test Chrome 39 explicitly. I confirmed
that 38 and 39 are currently failing in master, so I don't know
which version Sauce is giving us when we don't specify the version. It
seems pointless to test it without the version, if we aren't actually
getting the latest stable.
