+++ lib/adapters/idb/idb.js
@@ -6,6 +6,7 @@ var errors = require('../../deps/errors');
 var idbUtils = require('./idb-utils');
 var idbConstants = require('./idb-constants');
 var idbBulkDocs = require('./idb-bulk-docs');
+var checkBlobSupport = require('./idb-blob-support');
 
 var ADAPTER_VERSION = idbConstants.ADAPTER_VERSION;
 var ATTACH_AND_SEQ_STORE = idbConstants.ATTACH_AND_SEQ_STORE;
@@ -997,52 +998,9 @@ function init(api, opts, callback) {
         };
       }
 
-      // Detect blob support. Chrome didn't support it until version 38.
-      // in version 37 they had a broken version where PNGs (and possibly
-      // other binary types) aren't stored correctly.
       if (!blobSupportPromise) {
-
-        // make sure blob support is only checked one
-        blobSupportPromise = new utils.Promise(function (resolve, reject) {
-          // 1x1 transparent PNG
-          var blob = utils.createBlob([utils.fixBinary(utils.atob(
-            'iVBORw0KGgoAAAANSUhEUgAAAAEAAAA' +
-            'BCAQAAAC1HAwCAAAAC0lEQVQYV2NgYA' +
-            'AAAAMAAWgmWQ0AAAAASUVORK5CYII='
-          ))], {type: 'image/png'});
-          txn.objectStore(DETECT_BLOB_SUPPORT_STORE).put(blob, 'key');
-          txn.oncomplete = function () {
-            // have to do it in a separate transaction, else the correct
-            // content type is always returned
-            var blobTxn = idb.transaction([DETECT_BLOB_SUPPORT_STORE],
-              'readwrite');
-            var getBlobReq = blobTxn.objectStore(
-              DETECT_BLOB_SUPPORT_STORE).get('key');
-            getBlobReq.onerror = reject;
-            getBlobReq.onsuccess = function (e) {
-
-              var storedBlob = e.target.result;
-              var url = URL.createObjectURL(storedBlob);
-
-              utils.ajax({
-                url: url,
-                cache: true,
-                binary: true
-              }, function (err, res) {
-                if (err && err.status === 405) {
-                  // firefox won't let us do that. but firefox doesn't
-                  // have the blob type bug that Chrome does, so that's ok
-                  resolve(true);
-                } else {
-                  resolve(!!(res && res.type === 'image/png'));
-                }
-                URL.revokeObjectURL(url);
-              });
-            };
-          };
-        }).catch(function () {
-          return false; // error, so assume unsupported
-        });
+        // make sure blob support is only checked once
+        blobSupportPromise = checkBlobSupport(txn, idb);
       }
 
       blobSupportPromise.then(function (val) {
